feat: Add mydata command to provide information about the collection of user data
This commit adds a new command called "mydata" to the utility module. The command provides information about the collection of information about the user. It includes an embed with a description of the data collection process and a link to the detailed documentation on GitHub. The embed also displays the bot version and the bot's dynamic avatar URL.