feat: Add botUpdateCheck.js for checking bot updates
This commit adds a new file, botUpdateCheck.js, which contains a function for checking for updates to the bot. The function downloads the latest version of the package.json file from the GitHub repository and compares the version numbers. If a new version is available, the user is prompted to update the bot. If the user chooses to update, the bot is updated accordingly. This function should be called in client.js.